+++ app/src/main/java/com/mg/barpos/data/MenuItem.kt
@@ -2,7 +2,7 @@ package com.mg.barpos.data
 
 data class MenuItem(
     val itemName: String,
-    val itemPrice: Int,
+    val itemPrice: Double,
     val hasSides: Boolean,
     val sideOptions: Array<String>,
     val selectedSides: Array<String>,
